+++ src/main/resources/lessons/securepasswords/documentation/SecurePasswords_3.adoc
