+++ src/main/resources/lessons/securepasswords/documentation/SecurePasswords_4.adoc
